Fixing a logic error
I changed it such that the encoder will actually tell you the distance moved since the last time it was called, not the total distance moved as this would be problematic if you were using two different commands during that time.
I have also used two different constants,
distancePerRotation meaning the circumference of the motor.
movementPerRotation meaning the actual distance travelled by the robot when the move motors rotate once.